object Object [#41]]#43][#44][Close #38]. All commands except for the previously disabled ~active should now be back to their full functionality. The API, commands and data structures have all been moved outside of stats.js in to their new respective files as per #38 to take advantage of the new DBO module loader.
