+++ strings.json
@@ -6,10 +6,10 @@
         "english": "I've read {lines} messages in {chan} since {start}"
     },
     "user_words": {
-        "english": "{user} has spoken {words} words ({avg} wpl) in {chan} since {start}"
+        "english": "{user} has spoken {words} words ({avg}) in {chan} since {start}"
     },
     "chan_words": {
-        "english": "I've read {words} words ({avg} wpl) in {chan} since {start}"
+        "english": "I've read {words} words ({avg}) in {chan} since {start}"
     },
     "no_data": {
         "english": "I don't recall seeing {user} in here, are you sure they exist?"
@@ -43,5 +43,8 @@
     },
     "user_last": {
       "english": "{user} last seen in {chan} on {last}"
+    },
+    "chan_last": {
+      "english": "Last activity in {chan} recorded on {last}"
     }
 }
